pyproject.toml
[build-system]
requires = ["setuptools==80.9.0", "wheel==0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "homeprotect-dash"
version = "0.2.0"
description = "Agentic Homeprotect review intelligence dashboard built with Dash and Google ADK patterns"
readme = "README.md"
requires-python = ">=3.13"
authors = [
  { name = "Stefanos Cunning" }
]
dependencies = [
  "dash==3.2.0",
  "plotly==6.3.0",
  "pandas==2.3.2",
  "pydantic==2.11.7",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.2.2",
  "google-adk==1.26.0"
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.1",
  "pytest-cov==6.2.1",
  "ruff==0.12.7",
  "mypy==1.16.1"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
